Centralise bootstrap fixtures
We need these in the CLI tests
